add type directed value printer
